refactor blueprints (1/6): extract tasks blueprint
Five /tasks/* routes plus the post_task_status helper move to
blueprints/tasks.py. The blueprint is registered in app.py at the
bottom (`from blueprints.tasks import tasks_bp; app.register_blueprint(tasks_bp)`).
New blueprints/ package created with empty __init__.py.

Routes moved:
- POST /tasks/add (legacy public, JSON-backed, fires status update + omg.lol mirror)
- POST /tasks/complete (legacy public)
- POST /tasks/delete (legacy public)
- POST /tasks/<id>/edit (per-id, SQLite-backed)
- POST /tasks/<id>/assign (per-id, SQLite-backed)

Internal helper:
- post_task_status — only caller is /tasks/add, blueprint-internal

Smoke-tested: /publish 200 (cross-blueprint), /below-deck 200
(cross-blueprint), POST /tasks/999/edit and /tasks/999/assign return
404 (route resolves, auth passes, SQL runs, task not found — correct
for nonexistent ID).

First sub-phase of 6 in the blueprints split. Pattern proven —
remaining blueprints follow the same cadence.
